chore: move useAuth and useSocket hooks to src/hooks folder
